[Infra] Split refcount clang-tidy checks
Move refcount lifecycle analysis out of the general style checks and into a dedicated clang-tidy module with its own fixture coverage. Refcounting is an ownership contract anchored by object layout, so it needs a separate home from naming and formatting policy as the model grows.

Document the offset-zero refcount anchor convention, keep the VM counter exception explicit, and wire the dedicated refcount test through the Bazel and CMake plugin projects.
